pyproject.toml
[project]
name = "parasight"
version = "0.1.0"
description = "ParaSight, an AI driven, vision first UI test runner that latches onto your app’s front end and refuses to break, out‑adapting Selenium and Playwright at every DOM twist."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.14",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
packages = ["parasight"]
package-dir = {"" = "src"}

[tool.hatch.version]
path = "src/parasight/__version__.py"

[project.scripts]
parasight = "parasight.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/parasight"]

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default
lint.select = ["E", "F", "I"]
# Allow unused variables when underscore-prefixed
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
lint.ignore = ["E272"]
# Python version compatibility
target-version = "py312"
# Line length matches black's default
line-length = 120

[tool.ruff.format]
preview = true

[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]

